Upload notebook file
Add .ipynb file containing the updated code (in
order to work with the new version of yfinance) to
produce a graph about portfolio efficiency.
This code is a version of the original one that has
been written by university professor (credits in
notebook) and optimized to work with the new update
of YahooFinance dataset.